// src/models/cartlist.ts
export interface Cartlist {
    id:number;
    title: string;
    productCode: string;
    brand: string;
    soldBy: string;
    price: number;
    originalPrice: number;
    offerPercent: number;
    gender: string;
    size: any;
    description: string[];
    imgSrc: string;
    totalprice:number;
    quantity:number;
}